+++ flybywire-aircraft-a320-neo/html_ui/Pages/A32NX_Core/A32NX_ATSU.js
@@ -235,10 +235,11 @@ const insertUplink = (mcdu) => {
             CDUPerformancePage.UpdateEngOutAccFromOrigin(mcdu);
 
             await mcdu.tryUpdateAltDestination(alternateIcao);
-
+            SimVar.SetSimVarValue("L:A32NX_INITFLIGHT_AOC", "Number", 1);
             setTimeout(async () => {
                 await uplinkRoute(mcdu);
                 mcdu.addNewMessage(NXSystemMessages.aocActFplnUplink);
+                SimVar.SetSimVarValue("L:A32NX_INITFLIGHT_AOC", "Number", 2);
             }, mcdu.getDelayRouteChange());
 
             if (mcdu.page.Current === mcdu.page.InitPageA) {
@@ -246,17 +247,20 @@ const insertUplink = (mcdu) => {
             }
         }
     });
+    SimVar.SetSimVarValue("L:A32NX_INITFLIGHT_FLTNBR", "Number", 1);
     mcdu.updateFlightNo(fltNbr, (result) => {
         if (result) {
             if (mcdu.page.Current === mcdu.page.InitPageA) {
                 CDUInitPage.ShowPage1(mcdu);
             }
         }
+        SimVar.SetSimVarValue("L:A32NX_INITFLIGHT_FLTNBR", "Number", 2);
     });
 
     /**
      * INIT PAGE DATA UPLINK
     */
+    SimVar.SetSimVarValue("L:A32NX_INITFLIGHT_UPLINK", "Number", 1);
     setTimeout(() => {
         mcdu.setCruiseFlightLevelAndTemperature(cruiseAltitude);
         mcdu.tryUpdateCostIndex(costIndex);
@@ -264,6 +268,7 @@ const insertUplink = (mcdu) => {
         if (mcdu.page.Current === mcdu.page.InitPageA) {
             CDUInitPage.ShowPage1(mcdu);
         }
+        SimVar.SetSimVarValue("L:A32NX_INITFLIGHT_UPLINK", "Number", 2);
     }, mcdu.getDelayHigh());
 };